Added Python for FLUTE (for fast RSMT in HFNS).
* New: In flute, added a Python binding. Contains two methods: "readLUT()",
    to load the POWV9.dat and POST9.dat and "flute()" to build a RSMT.
    "flute()" takes a tuple of positions (themselves 2-uple) like:
        ( (x0,y0), (x1,y1,), ... (xN,yM) )
    and returns a tuple of 3-uple of branches:
        ( (n0,x0,y0), (n1,x1,y1), ... )
    In "flute.h", set the distance type (DTYPE) to int64_t to always
    accomodate DbU::Unit.
      As it now uses the Hurricane Python interface and the path utilities
    from CRL Core, move it's compilation *after* them (see build.conf in
    bootstrap).

+++ katana/src/GlobalRoute.cpp
@@ -446,9 +446,9 @@ namespace Katana {
         updateEstimateDensityOfPath( this, targets[0], targets[1], weight );
         return;
       default:
-        { int  accuracy = 3;
-          int* xs       = new int [targets.size()];
-          int* ys       = new int [targets.size()];
+        { int      accuracy = 3;
+          int64_t* xs       = new int64_t [targets.size()];
+          int64_t* ys       = new int64_t [targets.size()];
 
           for ( size_t itarget=0 ; itarget<targets.size() ; ++itarget ) {
             Point center =  targets[itarget]->getCenter();
